Add files via upload
Port Scanner - This scanner, written in Python, scans all ports within a specified range and on a given host address. This is a simple and beginner-friendly project that I decided on doing. I did have help from outside resources to create this code, as I was still in the intermediate phase of learning Python and wanted to learn while working on new projects. This project should show the viewer my proficiency in Python and my willingness to learn new things by taking the initiative.